+++ README.md
@@ -1,7 +1,6 @@
 [![Build Status](https://travis-ci.org/jhprinz/msm-tis.svg?branch=visualization)](https://travis-ci.org/jhprinz/msm-tis)
 
-opentis
-=======
+# opentis
 
 An open source Python framework for transition interface and path sampling calculations.
