--- docs/README.rst
@@ -1,2 +1,0 @@
-.. _README:
-.. importmd:: opentis/README.md
