+++ docs/conf.py
@@ -61,12 +61,15 @@ extensions = [
 #napoleon_use_param = True
 #napoleon_use_rtype = True
 
+# add pandoc directives
+pandoc_from = ['markdown', 'mediawiki']
+
 autosummary_generate = True
 autodoc_default_flags = ['members', 'inherited-members']
 
 sys.path.insert(0, os.path.abspath('sphinxext'))
 extensions.append('notebook_sphinxext')
-extensions.append('importmarkdown_sphinxext')
+extensions.append('pandoc_sphinxext')
 
 # Numpydoc options
 numpydoc_show_class_members = False
