+++ docs/netcdf_format.rst
@@ -1,11 +1,13 @@
 .. _netCDFFormatSpec:
 
-OpenPathSampling netCDF Format Specification
-============================================
+OPS File Format
+===============
 
 Overview
 --------
 
-Cool format. Allows independent frame storage positions, velocities, including lots of additional objects
-necessary. Easy export to mdtraj native objects possible so export to all sorts of file formats possible.
-Most objects are stored using human readable JSON String of Python objects.
+- Cool format.
+- Allows independent frame storage positions, velocities, including lots of additional objects
+necessary.
+- Easy export to mdtraj native objects possible so export to all sorts of file formats possible.
+- Most objects are stored using human readable JSON String of Python objects.
