+++ docs/orderparameter.rst
@@ -1,0 +1,40 @@
+.. _orderparameter:
+
+.. currentmodule:: opentis.orderparameter
+
+Orderparameter Functions
+========================
+
+:class:`opentis.Orderparameters`
+
+    >>> import openpathsampling as paths
+    >>> mover = paths.Orderparameter()
+
+basic orderparamters
+--------------------
+.. autosummary::
+    :toctree: api/generated/
+
+    OrderParameter
+    OP_Function
+
+simple orderparameters
+----------------------
+.. autosummary::
+    :toctree: api/generated/
+
+    OP_Volume
+
+
+mdtraj orderparamters
+---------------------
+.. autosummary::
+    :toctree: api/generated/
+
+    OP_MD_Function
+    OP_RMSD_To_Lambda
+
+msmbuilder3 orderparameters
+---------------------------
+
+    OP_Featurizer
