+++ docs/snapshot.rst
@@ -1,0 +1,21 @@
+.. _snapshot:
+
+.. currentmodule:: opentis
+
+Snapshot Functions
+==================
+
+:class:`opentis.Snapshot`
+
+    >>> import openpathsampling as paths
+    >>> mover = paths.PathMover()
+
+atomic movers
+-------------
+
+These are the leaves of a path tree. The atomic moved that can be made
+
+.. autosummary::
+    :toctree: api/generated/
+
+    Snapshot
