+++ docs/storage.rst
@@ -1,0 +1,28 @@
+.. _storage:
+
+.. currentmodule:: opentis.storage
+
+Storage Functions
+=================
+
+:class:`opentis.storage`
+
+    >>> import openpathsampling as paths
+    >>> mover = paths.storage.Storage()
+
+storages
+--------
+
+These are the leaves of a path tree. The atomic moved that can be made
+
+.. autosummary::
+    :toctree: api/generated/
+
+    Storage
+
+stores
+------
+.. autosummary::
+    :toctree: api/generated/
+
+    ObjectStore
